src/python/requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements/dev.txt requirements.in/dev.txt
#
--trusted-host pypi.org
--trusted-host files.pythonhosted.org
--trusted-host pypi.python.org

build==1.2.1
    # via pip-tools
chardet==5.2.0
    # via -r requirements.in/common.txt
click==8.1.7
    # via pip-tools
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via -r requirements.in/common.txt
gnupg==2.3.1
    # via -r requirements.in/common.txt
packaging==24.1
    # via build
pip-tools==7.4.1
    # via -r requirements.in/dev.txt
psutil==6.0.0
    # via gnupg
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
smmap==5.0.1
    # via gitdb
wheel==0.43.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
